requirements.txt
streamlit
spacy
nltk
python-docx
pdfplumber
langdetect
openai
